Add files via upload
Something fishy is going on with the serialization, JSON export of a test composition was 14KB but the same serialized was 8.62 MB.
Anyways I created the "Saver" interface so I don't need to rewrite the standard save code for JSON file export, audio export and whatever other export there is. Will probably also make a "Loader" interface for the inverse.